refactor: major modulerization
